refactor: rename page.tsx to index.tsx
