+++ src/shared/ui/NoticeDetail/index.tsx
@@ -1,22 +1,28 @@
-'use client'
+'use client';
 
 import { BadgeSelect } from '../BadgeSelect';
-import { Button } from '@/components/ui/button'
-import { Card, CardContent, CardFooter, CardHeader, CardTitle } from '@/components/ui/card'
-import { ArrowLeft, Calendar, Eye, User } from 'lucide-react'
-import { Notice } from '@/entities/notice/model/types'
+import { Button } from '@/components/ui/button';
+import {
+  Card,
+  CardContent,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+} from '@/components/ui/card';
+import { ArrowLeft, Calendar, Eye, User } from 'lucide-react';
+import { Notice } from '@/entities/notice/model/types';
 
 interface NoticeDetailProps {
-  notice: Notice
-  onBack: () => void
+  notice: Notice;
+  onBack: () => void;
 }
 
 export default function NoticeDetail({ notice, onBack }: NoticeDetailProps) {
   return (
-    <Card className="max-w-4xl mx-auto">
+    <Card className="mx-auto max-w-4xl">
       <CardHeader>
-        <div className="flex items-center justify-between mb-2">
-          <Button variant="ghost" onClick={onBack} className="p-0 h-8">
+        <div className="mb-2 flex items-center justify-between">
+          <Button variant="ghost" onClick={onBack} className="h-8 p-0">
             <ArrowLeft className="mr-2 h-4 w-4" />
             목록으로
           </Button>
@@ -32,7 +38,7 @@ export default function NoticeDetail({ notice, onBack }: NoticeDetailProps) {
           </div>
         </div>
         <CardTitle className="text-3xl font-bold">{notice.title}</CardTitle>
-        <div className="flex flex-wrap items-center gap-3 mt-3">
+        <div className="mt-3 flex flex-wrap items-center gap-3">
           <div className="flex items-center">
             <User className="mr-1 h-4 w-4 text-gray-500" />
             <span>{notice.author}</span>
@@ -44,15 +50,14 @@ export default function NoticeDetail({ notice, onBack }: NoticeDetailProps) {
       <CardContent>
         <div className="space-y-6">
           <div className="flex justify-center">
-            {notice.images.map((image) => (
+            {notice.images?.map((image) => (
               <img
                 key={image}
                 src={image}
                 alt={notice.title}
-                className="rounded-lg max-h-[500px] object-contain"
+                className="max-h-[500px] rounded-lg object-contain"
               />
-            )
-            )}
+            ))}
           </div>
 
           <div className="prose prose-lg max-w-none whitespace-pre-wrap">
@@ -68,5 +73,5 @@ export default function NoticeDetail({ notice, onBack }: NoticeDetailProps) {
         </Button>
       </CardFooter>
     </Card>
-  )
+  );
 }
